% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bq_ctas.R
\name{bq_ctas}
\alias{bq_ctas}
\title{bq_ctas}
\usage{
bq_ctas(
  ctas_qry,
  projectId,
  datasetId,
  tableId,
  createDisposition = "CREATE_IF_NEEDED",
  writeDisposition = "WRITE_EMPTY"
)
}
\arguments{
\item{writeDisposition}{}
}
\description{
bq_ctas
}
\examples{
  bq_ctas(
    projectId = "cw-bquery",
    datasetId = "l_cw_data_science",
    tableId = "cw_leadercampus_hr_platform_monthly_podcast",
    writeDisposition = "WRITE_TRUNCATE")

  bq_ctas("cw-bquery",
   "p_visual",
   "ch_lda_article_dashboard_user_inspector_user_h1_model_uid_topics",
   writeDisposition = "WRITE_TRUNCATE")
}
